Convert opening variant analysis logs to a command
